avoid int overflow
git-svn-id: https://svn.apache.org/repos/asf/xmlbeans/trunk@1935062 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/xmlbeans/GDurationBuilder.java
@@ -322,7 +322,7 @@ public class GDurationBuilder implements GDurationSpecification, java.io.Seriali
      * modulo(a, b) = a - fQuotient(a,b)*b
      */
     private static int _mod(long a, int b, long quotient) {
-        return (int) (a - quotient * b);
+        return Math.toIntExact(a - quotient * b);
     }
